Added grunt-cli to the package.json file updated readme accordingly.
Refactored to work in CI with the plugin with SC integration.

Configured suite to run using npm.

White space removed.
