+++ package.json
@@ -4,17 +4,18 @@
   "private": true,
   "description": "Selenium Mocha Chai Saucelabs",
   "scripts": {
-    "test-selenium": "node_modules/.bin/mocha --timeout 120000 --reporter spec --require test/mocha-setup.js test/selenium/*.js"
+    "test": "node_modules/.bin/grunt"
   },
   "devDependencies": {
     "chai": "1.9.1",
     "grunt": "0.4.5",
     "grunt-mocha-parallel": "^0.1.7",
     "grunt-parallel": "^0.4.1",
+    "grunt-cli": "^0.1.13",
     "mocha": "1.18.2",
-    "saucelabs": "0.1.1",
+    "saucelabs": "^1.0.1",
     "selenium-standalone": "2.42.0-2.9.0-1",
-    "selenium-webdriver": "2.42.1"
+    "selenium-webdriver": "^2.48.2"
   },
   "engines": {
     "node": "0.10.29"
